agent/requirements.txt
# Mira — Python Dependencies
# Install: pip install -r requirements.txt

# Phase 2: Core
anthropic>=0.39.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0

# Phase 3: Memory
chromadb>=0.5.0
networkx>=3.2
sentence-transformers>=3.0.0

# Phase 4: Audio
openai-whisper>=20231117

# Phase 5: Computer use
Pillow>=10.0.0
pyautogui>=0.9.54

# Phase 6: PA module
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0

# Phase 7+: Trading & crypto
schedule>=1.2.0
cryptography>=41.0.0

# Phase 11: Dashboard API
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0

# Phase 11: Voice
httpx>=0.27.0

# Dev tooling
watchdog>=4.0.0
